+++ common/src/main/scala/io/apex/f1/data/Retirement.scala
@@ -1,6 +1,6 @@
 package io.apex.f1.data
 
-case class Retirement(vehicleIdx: Short) {
+case class Retirement(vehicleIdx: Int) {
 //  def fill(buffer: ByteBuf): Retirement = {
 //    val vehicleIdx = buffer.readUnsignedByte()
 //    Retirement(vehicleIdx)
